Converted Permission bitsets from Long to Int

+++ src/main/kotlin/com/serebit/diskord/Typealiases.kt
@@ -2,5 +2,5 @@ package com.serebit.diskord
 
 internal typealias Snowflake = Long
 internal typealias IsoTimestamp = String
-internal typealias BitSet = Long
+internal typealias BitSet = Int
 internal typealias UnixTimestamp = Long
